+++ changelog.md
@@ -1,7 +1,10 @@
 # v2.3.0
 
 ### Build 201
-- Updated licenses
+
+- iPadOS: Shows sidebar on launch in portrait mode. 
+
+- Updated licenses.
 
 The app will do a full resync on first launch. If the app crashes for you, please reset the app from the Settings app (or delete and install again) and setup again.